refactor: update imports to use interfaces for Cart and User types for better code organization, introduced handleApiError
