+++ src/interfaces/IFavorite.ts
@@ -11,3 +11,12 @@ export type IFavoriteItem = {
 
   amount?: number;
 };
+
+export interface IFavoriteContext {
+  favoriteItems: IFavoriteItem[];
+  handleAddToFavorites: (clickedItem: IFavoriteItem) => void;
+  getTotalItems: (items: IFavoriteItem[]) => number;
+  removeItemFromFavorites: (product_id: string) => void;
+  totalItemsCount: number;
+}
+
