+++ src/services/address-service.ts
@@ -1,4 +1,5 @@
 import api from "./api";
+import { handleApiError } from "@/errors/errorHandler";
 import { IAddress } from "../interfaces/IDbUser";
 
 const baseUrl = "/addresses";
@@ -20,26 +21,46 @@ export const addressService: AddressService = {
 };
 
 async function create(data: Omit<IAddress, "address_id" | "createdAt" | "updatedAt">): Promise<IAddress> {
-  const response = await api.post<IAddress>(baseUrl, data);
-  return response.data;
+  try {
+    const response = await api.post<IAddress>(baseUrl, data);
+    return response.data;
+  } catch (error) {
+    throw handleApiError(error);
+  }
 }
 
 async function findAll(user_id: string): Promise<IAddress[]> {
-  const response = await api.get<IAddress[]>(`${baseUrl}/user/${user_id}`);
-  return response.data;
+  try {
+    const response = await api.get<IAddress[]>(`${baseUrl}/user/${user_id}`);
+    return response.data;
+  } catch (error) {
+    throw handleApiError(error);
+  }
 }
 
 async function findOne(address_id: string): Promise<IAddress> {
-  const response = await api.get<IAddress>(`${baseUrl}/${address_id}`);
-  return response.data;
+  try {
+    const response = await api.get<IAddress>(`${baseUrl}/${address_id}`);
+    return response.data;
+  } catch (error) {
+    throw handleApiError(error);
+  }
 }
 
 async function update(address_id: string, data: Partial<IAddress>): Promise<IAddress> {
-  const response = await api.patch<IAddress>(`${baseUrl}/${address_id}`, data);
-  return response.data;
+  try {
+    const response = await api.patch<IAddress>(`${baseUrl}/${address_id}`, data);
+    return response.data;
+  } catch (error) {
+    throw handleApiError(error);
+  }
 }
 
 async function remove(address_id: string): Promise<void> {
-  await api.delete(`${baseUrl}/${address_id}`);
+  try {
+    await api.delete(`${baseUrl}/${address_id}`);
+  } catch (error) {
+    throw handleApiError(error);
+  }
 }
