--- src/services/auth/types.ts
@@ -1,12 +1,0 @@
-import { Role } from "../../interfaces";
-
-export type User = {
-  name?: string;
-  lastName?: string;
-  email?: string;
-  role?: Role;
-  user_id?: string;
-  isSignedIn?: boolean;
-  userStatus?: string;
-};
-
