--- src/services/favorite/index.ts
@@ -1,2 +1,0 @@
-export * from "./types";
-
